servers/grepx-prefect-server/requirements.txt
# Prefect Server Requirements
# ==========================

# Shared models
-e ../grepx-task-generator-server

# Core Prefect dependencies
prefect>=3.0.0

# Configuration
omegaconf>=2.3.0

# Database connections (for database interactions)
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.7
aiosqlite>=0.19.0
motor>=3.3.0  # For MongoDB

# HTTP and networking
httpx>=0.25.0
requests>=2.31.0

# Utilities
pyyaml>=6.0
click>=8.0
typer>=0.9.0